Encoder test and WIP autonomous
