+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/BasicAuto.java
@@ -1,0 +1,8 @@
+package org.firstinspires.ftc.teamcode.autonomous;
+
+/**
+ * Created by Owen on 10/19/2018.
+ */
+
+public class BasicAuto {
+}
