+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleop/ServoToggle.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.teleop;
 
 import com.qualcomm.robotcore.hardware.Servo;
